requirements.txt
async_bing_client
websockets
pyyaml